angus-core/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-auth-httplib2==0.1.0
google-api-python-client==2.80.0
supabase>=1.0.3
requests==2.28.2
python-dotenv==1.0.0
schedule==1.2.0
openai>=1.13.3
pytube>=12.1.0
pydub>=0.25.1
uuid>=1.30
tblib>=1.7.0
crewai>=0.28.0
httpx==0.23.3

# Removed Coral Protocol dependencies
# langchain>=0.1.0
# langchain-openai>=0.1.0
# langchain-core>=0.3.36
# langchain-community>=0.1.0
# langchain_experimental>=0.0.42
# langchain-mcp-adapters==0.0.3
# sseclient-py>=1.7.2
